@@ -21,8 +21,8 @@ android {
         }
     }
 
-    compileSdkVersion 27
-    buildToolsVersion '27.0.3'
+    compileSdkVersion 28
+    buildToolsVersion '28.0.3'
 
     defaultConfig {
         applicationId "de.stephanlindauer.criticalmaps"
@@ -55,6 +55,7 @@ android {
     lintOptions {
         warning 'MissingTranslation'
         warning 'StringFormatInvalid'
+        warning 'ExpiringTargetSdkVersion'
 
         // https://github.com/square/okio/issues/58
         // http://jakewharton.github.io/butterknife/
@@ -94,8 +95,10 @@ dependencies {
     compileOnly 'javax.annotation:jsr250-api:1.0'
     annotationProcessor "com.google.dagger:dagger-compiler:$dagger_version"
 
+    errorprone("com.google.errorprone:error_prone_core:2.3.1")
+
     testImplementation 'junit:junit:4.12'
-    testImplementation 'com.google.truth:truth:0.40'
+    testImplementation 'com.google.truth:truth:0.41'
     testImplementation "com.android.support:support-annotations:$support_version"
-    testImplementation 'org.mockito:mockito-core:2.15.0'
+    testImplementation 'org.mockito:mockito-core:2.19.0'
 }
