@@ -13,7 +13,7 @@ public class OwnLocationModelTest {
     public void hasPreciseLocation_impreciseLocationIsReportedAsImprecise() throws Exception {
         OwnLocationModel tested = new OwnLocationModel();
 
-        tested.setLocation(new GeoPoint(0d, 0d), 50.0f, 0L);
+        tested.setLocation(new GeoPoint(0d, 0d), 50.0f);
 
         assertThat(tested.hasPreciseLocation()).isFalse();
     }
@@ -22,7 +22,7 @@ public class OwnLocationModelTest {
     public void hasPreciseLocation_preciseLocationIsReportedAsPrecise() {
         OwnLocationModel tested = new OwnLocationModel();
 
-        tested.setLocation(new GeoPoint(0d, 0d), 49.9f, 0L);
+        tested.setLocation(new GeoPoint(0d, 0d), 49.9f);
 
         assertThat(tested.hasPreciseLocation()).isTrue();
     }
@@ -39,27 +39,7 @@ public class OwnLocationModelTest {
         OwnLocationModel tested = new OwnLocationModel();
 
         String expected = "{\"latitude\":\"40741895\",\"longitude\":\"-73989308\"}";
-        tested.setLocation(new GeoPoint(40.741895d, -73.989308d), 1.1f, 0L);
+        tested.setLocation(new GeoPoint(40.741895d, -73.989308d), 1.1f);
         assertThat(tested.getLocationJson().toString()).isEqualTo(expected);
-    }
-
-    @Test
-    public void isLocationFresh_oldLocationReturnsFalse() {
-        OwnLocationModel tested = new OwnLocationModel();
-        long timestamp = new Date().getTime() - OwnLocationModel.MAX_LOCATION_AGE - 1000;
-
-        tested.setLocation(new GeoPoint(54.321d, 43.210d), 1.1f, timestamp);
-
-        assertThat(tested.isLocationFresh()).isFalse();
-    }
-
-    @Test
-    public void isLocationFresh_newLocationReturnsTrue() {
-        OwnLocationModel tested = new OwnLocationModel();
-        long timestamp = new Date().getTime();
-
-        tested.setLocation(new GeoPoint(54.321d, 43.210d), 1.1f, timestamp);
-
-        assertThat(tested.isLocationFresh()).isTrue();
     }
 }
